Skip Pattern to exclude certains URLs (#14)
* Skip Pattern to exclude certains URLs

* review fixes

* Use mockito spy and not waits

* remove junit from children poms

+++ opentracing-spring-web/src/main/java/io/opentracing/contrib/spring/web/client/RestTemplateSpanDecorator.java
@@ -69,7 +69,7 @@ public interface RestTemplateSpanDecorator {
             Tags.HTTP_METHOD.set(span, request.getMethod().toString());
 
             if (request.getURI().getPort() != -1) {
-                Tags.PEER_PORT.set(span, (short)request.getURI().getPort());
+                Tags.PEER_PORT.set(span, request.getURI().getPort());
             }
         }
